Rename fee validator to ValidateMembershipFeeUseCase
