+++ tests/Unit/UseCases/ApplyForMembership/MembershipApplicationValidatorTest.php
@@ -10,18 +10,18 @@ use WMDE\Fundraising\MembershipContext\Tests\Fixtures\SucceedingEmailValidator;
 use WMDE\Fundraising\MembershipContext\UseCases\ApplyForMembership\ApplicationValidationResult as Result;
 use WMDE\Fundraising\MembershipContext\UseCases\ApplyForMembership\ApplyForMembershipRequest;
 use WMDE\Fundraising\MembershipContext\UseCases\ApplyForMembership\MembershipApplicationValidator;
+use WMDE\Fundraising\MembershipContext\UseCases\ValidateMembershipFee\ValidateMembershipFeeUseCase;
 use WMDE\Fundraising\PaymentContext\Domain\BankDataValidationResult as BankResult;
 use WMDE\Fundraising\PaymentContext\Domain\BankDataValidator;
 use WMDE\Fundraising\PaymentContext\Domain\IbanBlocklist;
 use WMDE\Fundraising\PaymentContext\Domain\Model\BankData;
 use WMDE\Fundraising\PaymentContext\Domain\Model\Iban;
 use WMDE\FunValidators\ConstraintViolation;
-use WMDE\FunValidators\Validators\EmailValidator;
-use WMDE\Fundraising\MembershipContext\UseCases\ApplyForMembership\MembershipFeeValidator;
 use WMDE\FunValidators\ValidationResult;
+use WMDE\FunValidators\Validators\EmailValidator;
 
 /**
- * @covers \WMDE\Fundraising\MembershipContext\UseCases\ApplyForMembership\MembershipFeeValidator
+ * @covers \WMDE\Fundraising\MembershipContext\UseCases\ValidateMembershipFee\ValidateMembershipFeeUseCase
  *
  * @license GNU GPL v2+
  * @author Kai Nissen < kai.nissen@wikimedia.de >
@@ -88,8 +88,8 @@ class MembershipApplicationValidatorTest extends \PHPUnit\Framework\TestCase {
 		return new IbanBlocklist( [] );
 	}
 
-	private function newFailingFeeValidator(): MembershipFeeValidator {
-		$feeValidator = $this->getMockBuilder( MembershipFeeValidator::class )
+	private function newFailingFeeValidator(): ValidateMembershipFeeUseCase {
+		$feeValidator = $this->getMockBuilder( ValidateMembershipFeeUseCase::class )
 			->disableOriginalConstructor()->getMock();
 
 		$feeValidator->method( 'validate' )
@@ -98,8 +98,8 @@ class MembershipApplicationValidatorTest extends \PHPUnit\Framework\TestCase {
 		return $feeValidator;
 	}
 
-	private function newSucceedingFeeValidator(): MembershipFeeValidator {
-		$feeValidator = $this->getMockBuilder( MembershipFeeValidator::class )
+	private function newSucceedingFeeValidator(): ValidateMembershipFeeUseCase {
+		$feeValidator = $this->getMockBuilder( ValidateMembershipFeeUseCase::class )
 			->disableOriginalConstructor()->getMock();
 
 		$feeValidator->method( 'validate' )
